Removed noapprox and created Poly and LLS classes
After our last narcpack progress check, we decided to create classes
rather than functions for our approximation methods. That way we can
give the classes their own functions to take advantange of the
approximations.

Mostly as an example, I've started with a simple Poly class for
representing polynomials. This class is inherited by LLS, which
implements a linear least-squares fit.